refactor category flow
