+++ core/src/main/java/com/freemanpivo/chassi/domain/port/operations/GetCategories.java
