+++ datastore-h2/src/main/java/com/freemanpivo/chassi/h2/operations/SearchCategoryData.java
@@ -1,7 +1,9 @@
 package com.freemanpivo.chassi.h2.operations;
 
+import com.freemanpivo.chassi.domain.exception.BusinessException;
+import com.freemanpivo.chassi.domain.exception.ErrorMessageEnum;
 import com.freemanpivo.chassi.domain.model.Category;
-import com.freemanpivo.chassi.domain.port.command.SearchAllCategories;
+import com.freemanpivo.chassi.domain.port.operations.GetCategories;
 import com.freemanpivo.chassi.h2.mapper.CategoryEntityMapper;
 import com.freemanpivo.chassi.h2.repository.CategoryEntityRepository;
 import lombok.RequiredArgsConstructor;
@@ -12,17 +14,26 @@ import java.util.stream.Collectors;
 
 @Component
 @RequiredArgsConstructor
-public class SearchCategories implements SearchAllCategories {
+public class SearchCategories implements GetCategories {
 
     private final CategoryEntityRepository repository;
     private final CategoryEntityMapper mapper;
 
     @Override
-    public List<Category> getAllCategories() {
+    public List<Category> getCategories() {
         return repository
                 .findAll()
                 .stream()
                 .map(mapper::toModel)
                 .collect(Collectors.toList());
+    }
+
+    @Override
+    public Category getCategoryById(Long id) {
+        final var category = repository.findById(id);
+        if(category.isPresent())
+            return mapper.toModel(category.get());
+
+        throw new BusinessException(ErrorMessageEnum.E011.getCode(), ErrorMessageEnum.E011.getMessage());
     }
 }
